Add in a site structure document to detail the expected site structure.
This document is meant to be a starting guideline for the layout of
the website's navigational structure and details which relate to the
website navigational structure.

This document, structure.md, should assist in documenting the standards
for additional sumbissions, and in the standards for the review of such
submissions; however, it is not an absolute document.  Rather, it is a
starting point for the current practices.

Issue-6